move switch/dispute related stuff to switch.py
